+++ src/e2e/rpc/rpc.e2e.utils.ts
@@ -1,12 +1,16 @@
 import { Network } from '../../dto'
+import { e2eUtil } from '../e2e.util'
 
 export const RpcE2eUtils = {
-  initConfig: (network: Network) => ({
-    network,
-    retryCount: 5,
-    retryDelay: 2000,
-    apiKey: {
-      v2: process.env.V2_API_KEY,
-    },
-  }),
+  initConfig: (network: Network, apiKey?: string) => {
+    return {
+      network,
+      verbose: e2eUtil.isVerbose,
+      retryCount: 5,
+      retryDelay: 2000,
+      apiKey: {
+        v4: apiKey ?? process.env.V4_API_KEY_MAINNET,
+      },
+    }
+  },
 }