+++ src/e2e/rpc/tatum.rpc.eon.spec.ts
@@ -1,8 +1,0 @@
-import { Network } from '../../dto'
-import { EvmE2eUtils } from './evm.e2e.utils'
-
-describe('Eon', () => {
-  describe('mainnet', () => {
-    EvmE2eUtils.e2e({ network: Network.EON, chainId: 7332 })
-  })
-})