+++ src/e2e/rpc/tatum.rpc.optimism.spec.ts
@@ -1,8 +1,0 @@
-import { Network } from '../../service'
-import { EvmE2eUtils } from './evm.e2e.utils'
-
-describe.skip('Optimism', () => {
-  describe('mainnet', () => {
-    EvmE2eUtils.e2e({ network: Network.OPTIMISM, chainId: 10 })
-  })
-})