+++ src/service/rpc/generic/AbstractBatchRpc.ts
@@ -3,7 +3,7 @@ import { Container } from 'typedi'
 import { TatumConnector } from '../../../connector/tatum.connector'
 import { JsonRpcCall, JsonRpcResponse } from '../../../dto'
 import { AbstractRpcInterface } from '../../../dto/rpc/AbstractJsonRpcInterface'
-import { CONFIG } from '../../../util'
+import { CONFIG, Utils } from '../../../util'
 import { TatumConfig } from '../../tatum'
 
 export abstract class AbstractBatchRpc implements AbstractRpcInterface {
@@ -15,15 +15,8 @@ export abstract class AbstractBatchRpc implements AbstractRpcInterface {
     this.config = Container.of(this.id).get(CONFIG)
   }
 
-  protected getRpcNodeUrl(subPath?: string): string {
-    const { apiKey, rpc, network } = this.config
-    if (apiKey) {
-      const url =
-        rpc?.nodes?.[0].url ||
-        `https://api.tatum.io/v3/blockchain/node/${network}/${apiKey.v1 ? apiKey.v1 : apiKey.v2}`
-      return url.concat(subPath || '')
-    }
-    return rpc?.nodes?.[0].url || `https://api.tatum.io/v3/blockchain/node/${network}`.concat(subPath || '')
+  getRpcNodeUrl(subPath?: string): string {
+    return Utils.getV1RpcUrl(this.config, subPath)
   }
 
   rawRpcCall(body: JsonRpcCall): Promise<JsonRpcResponse<any>> {